Seperate config into its own package under internal
